+++ pkg/proofs/sigma/compose/sigor/cartesian.go
@@ -45,18 +45,6 @@ type WitnessCartesian[W0, W1 sigma.Witness] struct {
 	W1 W1
 }
 
-// Bytes returns the canonical byte representation of the composed witness.
-func (w *WitnessCartesian[W0, W1]) Bytes() []byte {
-	if w == nil {
-		return nil
-	}
-
-	var out = []byte{}
-	out = sliceutils.AppendLengthPrefixed(out, w.W0.Bytes())
-	out = sliceutils.AppendLengthPrefixed(out, w.W1.Bytes())
-	return out
-}
-
 var _ sigma.Witness = (*WitnessCartesian[sigma.Witness, sigma.Witness])(nil)
 
 // CommitmentCartesian represents a binary OR-composed commitment.
